perf: Fix accidental quadratic horizontal concat
