+++ crates/polars-core/src/functions.rs
@@ -2,48 +2,11 @@
 //!
 //! Functions that might be useful.
 //!
+pub use crate::frame::horizontal::concat_df_horizontal;
+#[cfg(feature = "diagonal_concat")]
 use crate::prelude::*;
 #[cfg(feature = "diagonal_concat")]
 use crate::utils::concat_df;
-
-/// Concat [`DataFrame`]s horizontally.
-/// Concat horizontally and extend with null values if lengths don't match
-pub fn concat_df_horizontal(dfs: &[DataFrame]) -> PolarsResult<DataFrame> {
-    let max_len = dfs
-        .iter()
-        .map(|df| df.height())
-        .max()
-        .ok_or_else(|| polars_err!(ComputeError: "cannot concat empty dataframes"))?;
-
-    let owned_df;
-
-    // if not all equal length, extend the DataFrame with nulls
-    let dfs = if !dfs.iter().all(|df| df.height() == max_len) {
-        owned_df = dfs
-            .iter()
-            .cloned()
-            .map(|mut df| {
-                if df.height() != max_len {
-                    let diff = max_len - df.height();
-                    df.columns
-                        .iter_mut()
-                        .for_each(|s| *s = s.extend_constant(AnyValue::Null, diff).unwrap());
-                }
-                df
-            })
-            .collect::<Vec<_>>();
-        owned_df.as_slice()
-    } else {
-        dfs
-    };
-
-    let mut first_df = dfs[0].clone();
-
-    for df in &dfs[1..] {
-        first_df.hstack_mut(df.get_columns())?;
-    }
-    Ok(first_df)
-}
 
 /// Concat [`DataFrame`]s diagonally.
 #[cfg(feature = "diagonal_concat")]
